+++ src/zidx.h
@@ -234,8 +234,8 @@ int zidx_index_init_advanced(zidx_index* index,
                              z_stream* z_stream_ptr, int initial_capacity,
                              int window_size, int compressed_data_buffer_size);
 int zidx_index_destroy(zidx_index* index);
-int zidx_read(zidx_index* index, unsigned char *buffer, size_t nbytes);
-int zidx_read_advanced(zidx_index* index, unsigned char *buffer,
+size_t zidx_read(zidx_index* index, unsigned char *buffer, size_t nbytes);
+size_t zidx_read_advanced(zidx_index* index, unsigned char *buffer,
                        size_t nbytes, zidx_block_callback block_callback,
                        void *callback_context);
 int zidx_seek(zidx_index* index, off_t offset, int whence);
